Structure Established, Close To Finish
General structure with events and various managers is working as intended, the last main thing to do is to actually go and call events in each of the spots where they are intended to be called. Just need to weed out bugs that might be hiding within the call stacks. Other than that the only thing is aesthetics and fine-tuning animation timings.