Bug-Squashing
General bug-squashing trying to just get things that were already working to work again. Chromatic aberration was messing up and that state machine was acting iffy. Unity was open for 12+ hours so it might be partially because of that but yea. Also reorganized folders so all the ones related to the current bit of work are in a central folder.