Load Ordering Sorted
Managed to solve a lot of issues with null refs right when the game is starting, but now the boss slides in animations and only does ranged attacks or the combo move. This commit is mainly just a bookmark that I have the load ordering sorted from what I can tell.